TabuSearch
This is one of the QAP (quadratic assignment problem) test problems of Nugent et al.

Twenty departments are to be distributed among twenty locations, five in each row (see below). The purpose is to
reduce expenses among the departments involved. The cost is equal to the product of the flow and the rectilinear distance between any given pair of departments. The distance and flow matrices are also provided